refactor: core testlogger integration.
* Refactor the code base to integrate with Spekt.TestLogger.
* Update version to 3.0.x
* Cleanup older business logic
