+++ .appveyor.yml
@@ -1,4 +1,4 @@
-version: '2.1.{build}'
+version: '3.0.{build}'
 image: Visual Studio 2019
 
 clone_depth: 1
