+++ NuGet.config
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <packageSources>
+    <clear />
+    <add key="github" value="https://nuget.pkg.github.com/spekt/index.json" />
+  </packageSources>
+  <packageSourceCredentials>
+    <github>
+      <add key="Username" value="codito" />
+      <add key="ClearTextPassword" value="de8f277eeab938891c108791ee53f49c4da8c717" />
+    </github>
+  </packageSourceCredentials>
+</configuration>
