+++ src/NUnit.Xml.TestLogger/NUnit.Xml.TestLogger.csproj
@@ -15,6 +15,10 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.TestPlatform.ObjectModel" Version="15.5.0" />
+    <PackageReference Include="Spekt.TestLogger" Version="$(TestLoggerVersion)" />
+    <!--<PackageReference Include="Spekt.TestLogger" Version="$(TestLoggerVersion)">-->
+      <!--<PrivateAssets>all</PrivateAssets>-->
+    <!--</PackageReference>-->
     <PackageReference Include="System.ValueTuple" Version="4.3.0" />
   </ItemGroup>
 </Project>
