+++ src/package/package.csproj
@@ -26,6 +26,7 @@
     <CopyOutputSymbolsToOutputDirectory>false</CopyOutputSymbolsToOutputDirectory>
     <CopyBuildOutputToPublishDirectory>false</CopyBuildOutputToPublishDirectory>
     <CopyOutputSymbolsToPublishDirectory>false</CopyOutputSymbolsToPublishDirectory>
+    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
 
     <!-- Disable warning that there are no source files. It is intentional. -->
     <NoWarn>$(NoWarn);2008;NU5127</NoWarn>
@@ -41,6 +42,10 @@
   <ItemGroup>
     <ProjectReference Include="..\NUnit.Xml.TestLogger\NUnit.Xml.TestLogger.csproj" />
     <ProjectReference Include="..\NUnit.Xml.TestLogger.TestAdapter\NUnit.Xml.TestLogger.TestAdapter.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Spekt.TestLogger" Version="$(TestLoggerVersion)" />
   </ItemGroup>
 
   <ItemGroup>
