+++ test/NUnit.Xml.TestLogger.UnitTests/NUnitXmlTestLoggerTests.cs
@@ -12,7 +12,7 @@ namespace NUnit.Xml.TestLogger.UnitTests
     using Microsoft.VisualStudio.TestPlatform.Extension.NUnit.Xml.TestLogger;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-    using TestSuite = Microsoft.VisualStudio.TestPlatform.Extension.NUnit.Xml.TestLogger.NUnitXmlTestLogger.TestSuite;
+    using TestSuite = Microsoft.VisualStudio.TestPlatform.Extension.NUnit.Xml.TestLogger.NUnitXmlSerializer.TestSuite;
 
     [TestClass]
     public class NUnitXmlTestLoggerTests
@@ -31,7 +31,7 @@ namespace NUnit.Xml.TestLogger.UnitTests
             var suite1 = CreateTestSuite("a.b");
             var suite2 = CreateTestSuite("c.d");
 
-            var result = NUnitXmlTestLogger.GroupTestSuites(new[] { suite1, suite2 }).ToArray();
+            var result = NUnitXmlSerializer.GroupTestSuites(new[] { suite1, suite2 }).ToArray();
 
             Assert.AreEqual(2, result.Length);
             Assert.AreEqual("a", result[0].Name);
@@ -45,7 +45,7 @@ namespace NUnit.Xml.TestLogger.UnitTests
             var expectedXmlForA = @"<test-suite type=""TestSuite"" name=""a"" fullname=""a"" total=""10"" passed=""2"" failed=""2"" inconclusive=""2"" skipped=""2"" result=""Failed"" duration=""0""><test-suite type=""TestSuite"" name=""b"" fullname=""a.b"" total=""10"" passed=""2"" failed=""2"" inconclusive=""2"" skipped=""2"" result=""Failed"" duration=""0""><test-suite /><test-suite /></test-suite></test-suite>";
             var expectedXmlForC = @"<test-suite type=""TestSuite"" name=""c"" fullname=""c"" total=""5"" passed=""1"" failed=""1"" inconclusive=""1"" skipped=""1"" result=""Failed"" duration=""0""><test-suite /></test-suite>";
 
-            var result = NUnitXmlTestLogger.GroupTestSuites(suites).ToArray();
+            var result = NUnitXmlSerializer.GroupTestSuites(suites).ToArray();
 
             Assert.AreEqual(2, result.Length);
             Assert.AreEqual("c", result[0].Name);
